move cube to box
